add LSTM, RNN
